chore!: Change safety of `Memory::read` to `unsafe`
As pointed out by [#10], `Memory::read` can cause undefined behaviour by
attempting to read an invalid bit pattern into a `T`. As such, unless we
further restrict the available implementations of `Memory` beyond
`T: Sized + Copy`.

I am currently of the opinion that restricting automatic implementations
of `Memory` should be avoided because it is not strictly the libary's
problem if the returned `T` is invalid. Instead, this function should be
marked as `unsafe`, with clear documentation for library consumers to
consider. If consumers would like to validate the bit patterns to ensure
defined behaviour, then they can do an intermediate load into an integer
or `[u8; N]` and check the bit pattern themselves before casting. This
is similar to what the standard library does for
[`std::mem::transmute`].

Resolves: #10
[#10]: https://github.com/Tommoa/rs-process-memory/issues/10
[`std::mem::transmute`]: https://doc.rust-lang.org/reference/behavior-considered-undefined.html